Configure default landing page.

+++ src/app/components/cart-app/cart-app.component.ts
@@ -1,11 +1,10 @@
 import { Component, OnInit } from '@angular/core';
-import { ProductService } from '../../services/product.service';
-import { Product } from '../../models/product';
 import { CatalogComponent } from '../catalog/catalog.component';
 import { CartItem } from '../../models/cartItem';
 import { NavbarComponent } from '../navbar/navbar.component';
-import { RouterOutlet } from '@angular/router';
+import { Router, RouterOutlet } from '@angular/router';
 import { SharingDataService } from '../../services/sharing-data.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'cart-app',
@@ -16,14 +15,14 @@ import { SharingDataService } from '../../services/sharing-data.service';
 })
 export class CartAppComponent implements OnInit{
 
-  products: Product[] = [];
   items: CartItem[] = [];
   total: number = 0;
 
-  constructor(private sharingDataService: SharingDataService, private service: ProductService){}
+  constructor(
+    private router: Router,
+    private sharingDataService: SharingDataService){}
   
   ngOnInit(): void {
-    this.products = this.service.findAll();
     //this.items = JSON.parse(sessionStorage.getItem('cart')!) || [];
     this.items = JSON.parse(sessionStorage.getItem('cart') || '[]');
     this.calculateTotal();
@@ -48,21 +47,56 @@ export class CartAppComponent implements OnInit{
         //this.items = [...this.items, {product, quantity:1}];
         this.items = [...this.items, {product: {...product}, quantity:1}];
       }
+
       this.calculateTotal();
       this.saveSession();
+      this.router.navigate(['/cart'],{
+        state: {items: this.items, total: this.total}
+      })
+
+      Swal.fire({
+        title: "Shopping Cart",
+        text: "New product added",
+        icon: "success"
+      });
     });
   }
 
   onDeleteCart(): void{
     this.sharingDataService.idProductEventEmitter.subscribe(id => {
       console.log(id + ' se ha ejecutado el evento idProductEventEmitter');
-      this.items = this.items.filter(item => item.product!.id !== id);
-      if(this.items.length == 0){
-        sessionStorage.removeItem('cart');
-        sessionStorage.clear();
-      }
-      this.calculateTotal();
-      this.saveSession();
+
+      Swal.fire({
+        title: "Are you sure to delete?",
+        text: "Care the product will be removed from the shopping cart.",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!"
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.items = this.items.filter(item => item.product!.id !== id);
+          if(this.items.length == 0){
+            sessionStorage.removeItem('cart');
+            sessionStorage.clear();
+          }
+          this.calculateTotal();
+          this.saveSession();
+    
+          this.router.navigateByUrl('/', {skipLocationChange: true}).then(() => {
+            this.router.navigate(['/cart'],{
+              state: {items: this.items, total: this.total}
+            });
+          });
+          
+          Swal.fire({
+            title: "Deleted!",
+            text: "The product has been removed from the shopping cart.",
+            icon: "success"
+          });
+        }
+      });
     })
   }
